requirements.txt
albumentations==1.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
imageio==2.23.0
joblib==1.2.0
networkx==2.8.8
numpy==1.24.1
opencv-python-headless==4.7.0.68
packaging==22.0
Pillow==9.4.0
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.13
